+++ jest.config.js
@@ -1,6 +1,0 @@
-export default {
-  testEnvironment: "jsdom",
-  testMatch: ["**/tests/unit/**/*.test.js"],
-  transform: {},
-  setupFiles: ["<rootDir>/jest.setup.js"],
-};